packages/meter/pyproject.toml
[project]
name = "meter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Marco Vinci", email = "m.vinci16@studenti.uniba.it" }]
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "paho-mqtt>=2.1.0",
    "polars>=1.39.3",
]

[project.scripts]
meter = "meter.main:main"

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"
